requirements.txt
asgiref==3.8.1
certifi==2024.2.2        
charset-normalizer==3.3.2
distlib==0.3.8
djando==0.2
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.1
filelock==3.14.0
idna==3.7
pillow==10.3.0
platformdirs==4.2.2
requests==2.32.2
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2